+++ chromium/chrome/browser/preloading/prefetch/search_prefetch/search_prefetch_service.cc
@@ -456,7 +456,7 @@ void SearchPrefetchService::OnPrerenderedRequestUsed(
   auto request_it = prefetches_.find(canonical_search_url);
   DCHECK(request_it != prefetches_.end());
   if (request_it == prefetches_.end()) {
-    // TODO(https://crbug.com/1295170): It should be rare but the request can be
+    // TODO(crbug.com/40214220): It should be rare but the request can be
     // deleted by timer before chrome activates the page. Add some metrics to
     // understand the possibility.
     return;
@@ -479,14 +479,14 @@ SearchPrefetchService::TakePrerenderFromMemoryCache(
   auto iter =
       RetrieveSearchTermsInMemoryCache(tentative_resource_request, recorder);
   if (iter == prefetches_.end()) {
-    // TODO(https://crbug.com/1414058): Recorder's state should not be
+    // TODO(crbug.com/40255935): Recorder's state should not be
     // kPrerendered, but it happened unexpectedly due to
     // restarting/serviceworker interception within prerender navigation stack
     // on ChromeOS.
     return {};
   }
 
-  // TODO(https://crbug.com/1295170): Do not use the prefetched response if it
+  // TODO(crbug.com/40214220): Do not use the prefetched response if it
   // is about to expire.
   DCHECK_NE(iter->second->current_status(),
             SearchPrefetchStatus::kRequestFailed);
@@ -646,7 +646,7 @@ void SearchPrefetchService::OnResultChanged(content::WebContents* web_contents,
     if (!prefetch_request->ShouldBeCancelledOnResultChanges()) {
       // Reset all pending prerenders. It will be set soon if service still
       // wants clients to prerender a SearchTerms.
-      // TODO(https://crbug.com/1295170): Unlike prefetch, which does not
+      // TODO(crbug.com/40214220): Unlike prefetch, which does not
       // discard completed response to avoid wasting, prerender would like
       // to cancel itself given the cost of a prerender. For now prenderer is
       // canceled when the prerender hints changed, we need to revisit this
@@ -691,7 +691,7 @@ void SearchPrefetchService::OnResultChanged(content::WebContents* web_contents,
     // comparatively higher than the prefetch to avoid the impact of wrong
     // predictions. We set confidence as 80 for prerender matches and 60 for
     // prefetch as an approximate number to differentiate both these cases.
-    int64_t confidence = BaseSearchProvider::ShouldPrerender(match) ? 80 : 60;
+    int confidence = BaseSearchProvider::ShouldPrerender(match) ? 80 : 60;
     auto* preloading_data =
         content::PreloadingData::GetOrCreateForWebContents(web_contents);
     SetIsNavigationInDomainCallback(preloading_data);
