+++ chromium/content/browser/cache_storage/cache_storage_manager_unittest.cc
@@ -122,7 +122,7 @@ class DelayedBlob : public storage::FakeBlob {
 
     // This should always succeed immediately because we size the pipe to
     // hold the entire blob for tiny data lengths.
-    uint32_t num_bytes = data_.length();
+    size_t num_bytes = data_.length();
     producer_handle_->WriteData(data_.data(), &num_bytes,
                                 MOJO_WRITE_DATA_FLAG_NONE);
     ASSERT_EQ(data_.length(), num_bytes);
@@ -1873,7 +1873,7 @@ TEST_F(CacheStorageManagerTest, GetBucketUsageConflictingBucketIds) {
   EXPECT_NE(Size(bucket_locator1_), 0);
 }
 
-// TODO(crbug.com/1369300): Re-enable test for Windows.
+// TODO(crbug.com/40868994): Re-enable test for Windows.
 #if BUILDFLAG(IS_WIN)
 #define MAYBE_GetBucketUsageWithPadding DISABLED_GetBucketUsageWithPadding
 #else
