+++ chromium/content/browser/download/mhtml_generation_manager.cc
@@ -12,6 +12,7 @@
 #include "base/functional/bind.h"
 #include "base/memory/ptr_util.h"
 #include "base/metrics/histogram_macros.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/stl_util.h"
 #include "base/strings/string_util.h"
 #include "base/strings/stringprintf.h"
@@ -220,7 +221,7 @@ class MHTMLGenerationManager::Job {
   // with on the fly hash computation.
   // Bound to the data pipe watcher and called upon notification of write
   // completion to producer pipe sent to the Renderer.
-  // TODO(https://crbug.com/915966): Eventually simplify this implementation
+  // TODO(crbug.com/40606905): Eventually simplify this implementation
   // with a DataPipeDrainer once error signalling is implemented there.
   void WriteMHTMLToDisk(MHTMLWriteCompleteCallback callback,
                         MojoResult result,
@@ -489,14 +490,15 @@ void MHTMLGenerationManager::Job::WriteMHTMLToDisk(
   DCHECK_NE(result, MOJO_RESULT_FAILED_PRECONDITION);
   // Begin consumer data pipe handle read and file write loop.
   char buffer[1024];
-  uint32_t num_bytes = sizeof(buffer);
+  size_t num_bytes = sizeof(buffer);
   while (result == MOJO_RESULT_OK && state.readable()) {
     result = mhtml_data_consumer_->ReadData(&buffer, &num_bytes,
                                             MOJO_READ_DATA_FLAG_NONE);
     if (result == MOJO_RESULT_OK) {
       if (secure_hash_)
         secure_hash_->Update(&buffer, num_bytes);
-      if (browser_file_.WriteAtCurrentPos(buffer, num_bytes) < 0) {
+      if (browser_file_.WriteAtCurrentPos(
+              buffer, base::checked_cast<int>(num_bytes)) < 0) {
         DLOG(ERROR) << "Error writing to file handle.";
         OnWriteComplete(std::move(callback),
                         mojom::MhtmlSaveStatus::kFileWritingError);
@@ -727,7 +729,7 @@ CloseFileResult MHTMLGenerationManager::Job::FinalizeOnFileThread(
                  "MHTMLGenerationManager::Job MHTML footer writing");
 
 #if BUILDFLAG(IS_FUCHSIA)
-    // TODO(crbug.com/1288816): Remove the Seek call.
+    // TODO(crbug.com/42050414): Remove the Seek call.
     // On fuchsia, fds do not share state. As the fd has been duped and sent to
     // the renderer process, it must be seeked to the end to ensure the data is
     // appended.
