+++ chromium/content/browser/service_worker/service_worker_cache_writer.cc
@@ -10,6 +10,7 @@
 #include "base/functional/bind.h"
 #include "base/memory/ptr_util.h"
 #include "base/memory/raw_ptr.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/task/sequenced_task_runner.h"
 #include "crypto/secure_hash.h"
 #include "crypto/sha2.h"
@@ -718,7 +719,7 @@ class ServiceWorkerCacheWriter::DataPipeReader {
             ReadCallback callback) {
     DCHECK(buffer);
     buffer_ = std::move(buffer);
-    num_bytes_to_read_ = num_bytes;
+    num_bytes_to_read_ = base::checked_cast<size_t>(num_bytes);
     callback_ = std::move(callback);
 
     if (!data_.is_valid()) {
@@ -747,10 +748,10 @@ class ServiceWorkerCacheWriter::DataPipeReader {
     if (result != MOJO_RESULT_OK) {
       // Disconnected means it's the end of the body or an error occurs during
       // reading the body.
-      // TODO(https://crbug.com/1055677): notify of errors.
+      // TODO(crbug.com/40120038): notify of errors.
       num_bytes_to_read_ = 0;
     }
-    owner_->AsyncDoLoop(num_bytes_to_read_);
+    owner_->AsyncDoLoop(base::checked_cast<int>(num_bytes_to_read_));
   }
 
   void OnReadDataPrepared(mojo::ScopedDataPipeConsumerHandle data) {
@@ -769,14 +770,14 @@ class ServiceWorkerCacheWriter::DataPipeReader {
                        weak_factory_.GetWeakPtr()));
     ReadInternal(MOJO_RESULT_OK);
 
-    // TODO(https://crbug.com/1055677): provide a callback to notify of errors
+    // TODO(crbug.com/40120038): provide a callback to notify of errors
     // if any.
     reader_->ReadData({});
   }
 
   // Parameters set on Read().
   scoped_refptr<net::IOBuffer> buffer_;
-  uint32_t num_bytes_to_read_ = 0;
+  size_t num_bytes_to_read_ = 0;
   ReadCallback callback_;
 
   // |reader_| is safe to be kept as a rawptr because |owner_| owns |this| and
