+++ chromium/content/browser/web_package/signed_exchange_cert_fetcher.cc
@@ -9,7 +9,6 @@
 #include "base/functional/bind.h"
 #include "base/metrics/histogram_macros.h"
 #include "base/numerics/safe_conversions.h"
-#include "base/strings/string_piece.h"
 #include "base/strings/stringprintf.h"
 #include "base/task/sequenced_task_runner.h"
 #include "base/task/single_thread_task_runner.h"
@@ -114,7 +113,7 @@ SignedExchangeCertFetcher::SignedExchangeCertFetcher(
       resource_request_(std::make_unique<network::ResourceRequest>()),
       callback_(std::move(callback)),
       devtools_proxy_(devtools_proxy) {
-  // TODO(https://crbug.com/803774): Revisit more ResourceRequest flags.
+  // TODO(crbug.com/40558902): Revisit more ResourceRequest flags.
   resource_request_->url = cert_url;
   // |request_initiator| is used for cookie checks, but cert requests don't use
   // cookies. So just set an opaque Origin.
@@ -186,7 +185,7 @@ void SignedExchangeCertFetcher::OnHandleReady(MojoResult result) {
   TRACE_EVENT0(TRACE_DISABLED_BY_DEFAULT("loading"),
                "SignedExchangeCertFetcher::OnHandleReady");
   const void* buffer = nullptr;
-  uint32_t num_bytes = 0;
+  size_t num_bytes = 0;
   MojoResult rv =
       body_->BeginReadData(&buffer, &num_bytes, MOJO_READ_DATA_FLAG_NONE);
   if (rv == MOJO_RESULT_OK) {
