+++ chromium/content/common/service_worker/race_network_request_write_buffer_manager.cc
@@ -80,7 +80,7 @@ void RaceNetworkRequestWriteBufferManager::CancelWatching() {
 
 MojoResult RaceNetworkRequestWriteBufferManager::BeginWriteData() {
   void* buffer;
-  uint32_t num_write_bytes;
+  size_t num_write_bytes;
   MojoResult result = producer_->BeginWriteData(&buffer, &num_write_bytes,
                                                 MOJO_WRITE_DATA_FLAG_NONE);
   buffer_ = base::make_span(static_cast<char*>(buffer), num_write_bytes);
@@ -106,7 +106,7 @@ std::tuple<MojoResult, size_t> RaceNetworkRequestWriteBufferManager::WriteData(
   auto buffer = read_buffer.size() > data_pipe_buffer_size_
                     ? read_buffer.subspan(0, data_pipe_buffer_size_)
                     : read_buffer;
-  uint32_t num_bytes = buffer.size();
+  size_t num_bytes = buffer.size();
   MojoResult result = producer_->WriteData(buffer.data(), &num_bytes,
                                            MOJO_WRITE_DATA_FLAG_ALL_OR_NONE);
   num_bytes_written_ += num_bytes;
@@ -137,7 +137,7 @@ size_t RaceNetworkRequestWriteBufferManager::CopyAndCompleteWriteDataWithSize(
   CHECK_GE(read_buffer.size(), num_bytes_to_consume);
   // Check if all memory spaces are available to access. `volatile` to avoid the
   // compiler optimization.
-  // TODO(crbug.com/1502946) Remove this code once we confirmed the root cause
+  // TODO(crbug.com/40943349) Remove this code once we confirmed the root cause
   // of the crash.
   volatile const char* read_buffer_v =
       static_cast<volatile const char*>(read_buffer.data());
